+++ R/zzz.R
@@ -1,6 +1,6 @@
 .onAttach <- function(...){
-  options("cheapr.cores" = getOption("cheapr.cores", 1),
-          "cheapr.digits" = getOption("cheapr.digits", 2))
+  set_threads(2)
+  options("cheapr.digits" = getOption("cheapr.digits", 2))
   fastplyr_pkg <- find.package("fastplyr", quiet = TRUE)
   if (length(fastplyr_pkg) > 0){
     fastplyr_version <- utils::packageVersion("fastplyr")