+++ tests/testthat.R
@@ -9,6 +9,6 @@
 library(testthat)
 library(cheapr)
 
-options(cheapr.cores = 2)
+set_threads(2)
 
 test_check("cheapr", reporter = c("check", "location"))