+++ NEWS.md
@@ -4,12 +4,15 @@ cheapr 1.5.0 supersedes cheapr 1.4.0
 
 - A new rich set of parallelised math functions like `abs_`, `round_` and more.
 
+- New functions for setting and getting the number of threads being used.
+
 - Plain list vectors are now never regarded as `NA` even if they contain `NA`
 elements.
 
 - Parallelisation is internally used more frequently for operations like 
 creating new vectors, combining vectors, filling, copying and replacing data.
-To use more than 1 thread/core, use `options(cheapr.cores = )`.
+The default has been changed to 2. To use a different number of threads simply
+use `set_threads()`
 
 - `scm` now better detects when integer (32-bit and 64-bit) overflow occurs and
 switches to using doubles internally, returning a double value.